Move block classes (#432)
